add basic persistence framework
This doesn't actually work at this point - spawning with initial
gear works, respawn works, but loading gear doesn't. This is because
it seems to be saved with nil, so savePlayer somehow doesn't quite
work, but I'm out of time and patience.

Also read the TODO in loadFromProfile; the saving logic might need
to be redesigned somewhat to account for server crashes and players
losing progress for that session (despite autosave saying that it
autosaved).

Signed-off-by: freghar <freghar@dummy.tld>